apps/backend/pyproject.toml
[project]
name = "birdie"
version = "0.1.0"
description = ""
authors = [
    {name = "Hojoong Jang",email = "hojoong310@gmail.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "fastapi (==0.115.12)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "py-ulid (>=1.0.3,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "bcrypt (<4.1.0)",
    "dependency-injector (==4.41.0)",
    "pydantic[email] (>=2.11.3,<3.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "h11 (==0.16.0)",
    "boto3 (>=1.38.13,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "portone-server-sdk (>=0.16.2,<0.17.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"